tracks/beginner_track/weak_rsa/requirements.txt
gmpy2==2.2.1
pycryptodome==3.20.0
